Factored out component subclass for Ember
