+++ apps/admin-x-framework/src/index.ts
@@ -1,5 +1,5 @@
 export {default as FrameworkProvider, useFramework} from './providers/FrameworkProvider';
-export type {FrameworkContextType, FrameworkProviderProps} from './providers/FrameworkProvider';
+export type {FrameworkContextType, FrameworkProviderProps, TopLevelFrameworkProps} from './providers/FrameworkProvider';
 
 export {useQueryClient} from '@tanstack/react-query';
 export type {InfiniteData} from '@tanstack/react-query';
