+++ apps/admin-x-framework/src/providers/FrameworkProvider.tsx
@@ -8,7 +8,7 @@ export interface FrameworkProviderProps {
     basePath: string;
     ghostVersion: string;
     externalNavigate: RoutingProviderProps['externalNavigate'];
-    modals: RoutingProviderProps['modals'];
+    modals?: RoutingProviderProps['modals'];
     unsplashConfig: {
         Authorization: string;
         'Accept-Version': string;
@@ -23,6 +23,9 @@ export interface FrameworkProviderProps {
 
     children: ReactNode;
 }
+
+// children, basePath and modals should be provided by each app, while others are passed in from Ember
+export type TopLevelFrameworkProps = Omit<FrameworkProviderProps, 'children' | 'basePath' | 'modals'>;
 
 export type FrameworkContextType = Omit<FrameworkProviderProps, 'basePath' | 'externalNavigate' | 'modals' | 'children'>;
