+++ apps/admin-x-settings/src/App.tsx
@@ -1,23 +1,23 @@
 import MainContent from './MainContent';
 import SettingsAppProvider, {OfficialTheme, UpgradeStatusType} from './components/providers/SettingsAppProvider';
 import SettingsRouter, {loadModals, modalPaths} from './components/providers/SettingsRouter';
-import {DesignSystemApp, FetchKoenigLexical} from '@tryghost/admin-x-design-system';
-import {FrameworkProvider, FrameworkProviderProps} from '@tryghost/admin-x-framework';
+import {DesignSystemApp, DesignSystemAppProps} from '@tryghost/admin-x-design-system';
+import {FrameworkProvider, TopLevelFrameworkProps} from '@tryghost/admin-x-framework';
 import {ZapierTemplate} from './components/settings/advanced/integrations/ZapierModal';
 
-interface AppProps extends Omit<FrameworkProviderProps, 'basePath' | 'modals' | 'children'> {
+interface AppProps {
+    framework: TopLevelFrameworkProps;
+    designSystem: DesignSystemAppProps;
     officialThemes: OfficialTheme[];
     zapierTemplates: ZapierTemplate[];
-    darkMode: boolean;
-    fetchKoenigLexical: FetchKoenigLexical;
     upgradeStatus?: UpgradeStatusType;
 }
 
-function App({officialThemes, zapierTemplates, upgradeStatus, darkMode, fetchKoenigLexical, ...props}: AppProps) {
+function App({framework, designSystem, officialThemes, zapierTemplates, upgradeStatus}: AppProps) {
     return (
-        <FrameworkProvider basePath='settings' modals={{paths: modalPaths, load: loadModals}} {...props}>
+        <FrameworkProvider basePath='settings' modals={{paths: modalPaths, load: loadModals}} {...framework}>
             <SettingsAppProvider officialThemes={officialThemes} upgradeStatus={upgradeStatus} zapierTemplates={zapierTemplates}>
-                <DesignSystemApp className='admin-x-settings' darkMode={darkMode} fetchKoenigLexical={fetchKoenigLexical} id='admin-x-settings'>
+                <DesignSystemApp className='admin-x-settings' {...designSystem}>
                     <SettingsRouter />
                     <MainContent />
                 </DesignSystemApp>
